exercise2/mysite/requirements.txt
Django==3.1.5
django-url-filter==0.3.15
djangorestframework==3.12.2
gunicorn==20.0.4
jupyterlab==3.0.5
psycopg2-binary==2.8.6
requests==2.25.1
